atualiza validacao de id e remove return

+++ src/controllers/clienteCtrl.js
@@ -29,8 +29,8 @@ class ClienteCtrl {
         try {
             const id = Number(req.params.id);
 
-            if(!id){
-                throw new AppError('ID inválido', 400)
+            if (isNaN(id) || id <= 0) {
+                throw new AppError('ID inválido', 400);
             }
 
             await ClienteService.deleterCliente(id);
